NewTube{,Register,Handler} now accept PriorityQueueCreator.
Prior to this refactoring, a hard-coded PriorityQueueCreator only
allowed MinHeaps to be used as a backend for beanstalkg.

This commit allows the user to choose their preferred backend, when
available.

When a user-specified backend is unavailable, beanstalkg chooses to
fallback to MinHeap instead of failing.

The new logic in config.go might be better placed in the backend
package.